# backend/src/openapi/wire.openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "WIRE API",
    "version": "1.0.0",
    "description": "REST API for wire transfer intent management, approvals, voice challenges, audit bundles, API keys, and webhooks.\n\nThis API provides:\n- Transfer intent lifecycle (create, update, execute)\n- Voice challenge and proof submission\n- Approval decisions and token management\n- Beneficiary management\n- Policy configuration and simulation\n- Event chain and audit bundle generation\n- API keys and webhooks (developer platform)\n\nAll endpoints are prefixed with `/api/wire`.\n\n**Authentication:**\n- UI sessions: `Authorization: Bearer <token>`\n- Integrations: `Authorization: Basic base64(<apiKey>:<apiSecret>)` (recommended) or `X-API-KEY: <apiKey>` (legacy)\n\n**Error Format:** All errors follow `{ error: string, code?: string, details?: object }` format.\n**Idempotency:** Mutating endpoints support `X-Idempotency-Key` (where applicable)."
  },
  "servers": [
    {
      "url": "http://localhost:3000/api/wire",
      "description": "Local development server"
    }
  ],
...
  "paths": {
...
    "/intents": {
      "get": {
        "tags": ["Intents"],
        "summary": "List transfer intents",
        "description": "Returns all transfer intents for the authenticated user's organization, ordered by creation date (newest first).",
        "operationId": "listIntents",
        "responses": {
          "200": {
            "description": "List of transfer intents",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TransferIntent"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Intents"],
        "summary": "Create transfer intent",
        "description": "Creates a new transfer intent. Requires `intent:create` permission.",
        "operationId": "createIntent",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateIntentRequest"
              },
              "examples": {
                "ach_transfer": {
                  "summary": "ACH transfer example",
                  "value": {
                    "railsType": "ACH",
                    "amountMinor": "50000",
                    "currency": "USD",
                    "beneficiaryId": "benef_123",
                    "purpose": "Payment for services rendered"
                  }
                },
                "wire_transfer": {
                  "summary": "International wire transfer example",
                  "value": {
                    "railsType": "WIRE",
                    "amountMinor": "1000000",
                    "currency": "USD",
                    "beneficiaryId": "benef_456",
                    "purpose": "International vendor payment"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Transfer intent created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransferIntent"
                },
                "examples": {
                  "created_intent": {
                    "summary": "Successfully created intent",
                    "value": {
                      "id": "intent_abc123",
                      "orgId": "org_123",
                      "createdByUserId": "user_456",
                      "railsType": "ACH",
                      "amountMinor": "50000",
                      "currency": "USD",
                      "beneficiaryId": "benef_123",
                      "beneficiaryVersion": 1,
                      "purpose": "Payment for services rendered",
                      "status": "DRAFT",
                      "riskScore": 25,
                      "riskRationaleJson": {
                        "factors": ["low_amount", "domestic"],
                        "details": {},
                        "scoreBreakdown": {
                          "base": 20,
                          "amount": 5,
                          "total": 25
                        }
                      },
                      "requiredApprovals": 1,
                      "requiredChallengeLevel": "L1",
                      "bindingHash": "abc123def456",
                      "cooldownUntil": null,
                      "createdAt": "2025-01-01T12:00:00Z",
                      "updatedAt": "2025-01-01T12:00:00Z"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request body",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "invalid_amount": {
                    "summary": "Invalid amount format",
                    "value": {
                      "error": "amountMinor must match pattern /^\\d+$/",
                      "code": "INVALID_REQUEST",
                      "details": {
                        "field": "amountMinor",
                        "received": "50.00",
                        "expected": "numeric string"
                      }
                    }
                  },
                  "missing_required": {
                    "summary": "Missing required field",
                    "value": {
                      "error": "beneficiaryId is required",
                      "code": "INVALID_REQUEST",
                      "details": {
                        "field": "beneficiaryId"
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Beneficiary not found or policy not configured",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "beneficiary_not_found": {
                    "summary": "Beneficiary not found",
                    "value": {
                      "error": "Beneficiary not found",
                      "code": "BENEFICIARY_NOT_FOUND",
                      "details": {
                        "beneficiaryId": "benef_999"
                      }
                    }
                  },
                  "policy_not_configured": {
                    "summary": "No policy configured",
                    "value": {
                      "error": "No policy configured",
                      "code": "POLICY_NOT_CONFIGURED"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/beneficiaries": {
      "get": {
        "tags": ["Beneficiaries"],
        "summary": "List beneficiaries",
        "description": "Returns all beneficiaries for the authenticated user's organization.",
        "operationId": "listBeneficiaries",
        "responses": {
          "200": {
            "description": "List of beneficiaries",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Beneficiary"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Beneficiaries"],
        "summary": "Create beneficiary",
        "description": "Creates a new beneficiary. Requires `beneficiary:create` permission.",
        "operationId": "createBeneficiary",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBeneficiaryRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Beneficiary created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Beneficiary"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request body",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api-keys": {
      "get": {
        "tags": ["API Keys"],
        "summary": "List API keys",
        "description": "Lists API keys for the authenticated organization. Requires `api:view` permission.",
        "operationId": "listApiKeys",
        "security": [
          { "bearerAuth": [] },
          { "basicAuth": [] },
          { "apiKeyHeader": [] }
        ],
        "responses": {
          "200": {
            "description": "List of API keys",
            "content": {
              "application/json": {
                "schema": { "type": "array", "items": { "type": "object" } }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["API Keys"],
        "summary": "Create API key",
        "description": "Creates a new API key. Returns `secret` only once. Requires `api:create` permission.",
        "operationId": "createApiKey",
        "security": [
          { "bearerAuth": [] }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "type": "object" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "API key created (secret returned once)",
            "content": {
              "application/json": {
                "schema": { "type": "object" }
              }
            }
          }
        }
      }
    },
    "/api-keys/{id}/revoke": {
      "post": {
        "tags": ["API Keys"],
        "summary": "Revoke API key",
        "description": "Revokes an API key. Requires `api:edit` permission.",
        "operationId": "revokeApiKey",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Revoked", "content": { "application/json": { "schema": { "type": "object" } } } },
          "404": { "description": "Not found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      }
    },
    "/api-keys/{id}": {
      "delete": {
        "tags": ["API Keys"],
        "summary": "Delete API key",
        "description": "Deletes an API key. Requires `api:delete` permission.",
        "operationId": "deleteApiKey",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Deleted", "content": { "application/json": { "schema": { "type": "object" } } } },
          "404": { "description": "Not found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "UI session auth (Bearer token)."
      },
      "basicAuth": {
        "type": "http",
        "scheme": "basic",
        "description": "Integration auth: base64(<apiKey>:<apiSecret>)"
      },
      "apiKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-KEY",
        "description": "Legacy integration auth (key-only). Prefer Basic auth."
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": {
            "type": "string",
            "description": "Human-readable error message"
          },
          "code": {
            "type": "string",
            "description": "Machine-parseable error code (optional)"
          },
          "details": {
            "type": "object",
            "description": "Additional error details (optional)"
          }
        }
      },
...
      "TransferIntent": {
        "type": "object",
        "required": [
          "id",
          "orgId",
          "createdByUserId",
          "railsType",
          "amountMinor",
          "currency",
          "beneficiaryId",
          "purpose",
          "status",
          "riskScore",
          "riskRationaleJson",
          "requiredApprovals",
          "requiredChallengeLevel",
          "bindingHash",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "id": {
            "type": "string",
            "description": "Transfer intent ID"
          },
          "orgId": {
            "type": "string",
            "description": "Organization ID"
          },
          "createdByUserId": {
            "type": "string",
            "description": "User ID who created the intent"
          },
          "railsType": {
            "type": "string",
            "enum": ["ACH", "WIRE"],
            "description": "Payment rails type"
          },
          "amountMinor": {
            "type": "string",
            "pattern": "^\\d+$",
            "description": "Amount in minor units (e.g., cents) as string"
          },
          "currency": {
            "type": "string",
            "minLength": 3,
            "maxLength": 3,
            "description": "ISO 4217 currency code (e.g., USD)"
          },
          "beneficiaryId": {
            "type": "string",
            "description": "Beneficiary ID"
          },
          "beneficiaryVersion": {
            "type": "integer",
            "description": "Beneficiary version snapshot (optional)"
          },
          "purpose": {
            "type": "string",
            "minLength": 1,
            "description": "Transfer purpose description"
          },
          "status": {
            "type": "string",
            "enum": [
              "DRAFT",
              "PENDING_PROOF",
              "CHALLENGING",
              "PENDING_APPROVALS",
              "APPROVED",
              "DENIED",
              "EXECUTED",
              "EXPIRED",
              "CANCELED"
            ],
            "description": "Intent status"
          },
          "riskScore": {
            "type": "integer",
            "minimum": 0,
            "maximum": 100,
            "description": "Risk score (0-100)"
          },
          "riskRationaleJson": {
            "$ref": "#/components/schemas/RiskRationale",
            "description": "Risk scoring rationale"
          },
          "requiredApprovals": {
            "type": "integer",
            "minimum": 1,
            "description": "Number of required approvals"
          },
          "requiredChallengeLevel": {
            "type": "string",
            "enum": ["L1", "L2", "L3"],
            "description": "Required voice challenge level"
          },
          "bindingHash": {
            "type": "string",
            "description": "Cryptographic hash of binding-critical fields"
          },
          "cooldownUntil": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "description": "Cooldown expiration timestamp (if applicable)"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Creation timestamp"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "Last update timestamp"
          }
        }
      },
      "RiskRationale": {
        "type": "object",
        "required": ["factors", "details", "scoreBreakdown"],
        "properties": {
          "factors": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "List of risk factors"
          },
          "details": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            },
            "description": "Detailed risk factor information"
          },
          "scoreBreakdown": {
            "type": "object",
            "required": ["base", "total"],
            "properties": {
              "base": {
                "type": "number",
                "description": "Base risk score"
              },
              "amount": {
                "type": "number",
                "description": "Amount-based risk adjustment"
              },
              "beneficiary": {
                "type": "number",
                "description": "Beneficiary-based risk adjustment"
              },
              "rails": {
                "type": "number",
                "description": "Rails-based risk adjustment"
              },
              "international": {
                "type": "number",
                "description": "International transfer risk adjustment"
              },
              "timing": {
                "type": "number",
                "description": "Timing-based risk adjustment"
              },
              "total": {
                "type": "number",
                "description": "Total risk score"
              }
            }
          }
        }
      },
      "CreateIntentRequest": {
        "type": "object",
        "required": ["railsType", "amountMinor", "beneficiaryId", "purpose"],
        "properties": {
          "railsType": {
            "type": "string",
            "enum": ["ACH", "WIRE"],
            "description": "Payment rails type"
          },
          "amountMinor": {
            "type": "string",
            "pattern": "^\\d+$",
            "description": "Amount in minor units (e.g., cents) as string"
          },
          "currency": {
            "type": "string",
            "minLength": 3,
            "maxLength": 3,
            "default": "USD",
            "description": "ISO 4217 currency code"
          },
          "beneficiaryId": {
            "type": "string",
            "minLength": 1,
            "description": "Beneficiary ID"
          },
          "purpose": {
            "type": "string",
            "minLength": 1,
            "description": "Transfer purpose description"
          }
        }
      },
...
      "Beneficiary": {
        "type": "object",
        "required": [
          "id",
          "orgId",
          "displayName",
          "country",
          "railsAllowed",
          "bankLast4",
          "bankTokenHash",
          "version",
          "status",
          "createdAt",
          "updatedAt",
          "lastChangedAt",
          "lastChangedBy"
        ],
        "properties": {
          "id": {
            "type": "string",
            "description": "Beneficiary ID"
          },
          "orgId": {
            "type": "string",
            "description": "Organization ID"
          },
          "displayName": {
            "type": "string",
            "description": "Display name"
          },
          "country": {
            "type": "string",
            "minLength": 2,
            "maxLength": 2,
            "description": "ISO 3166-1 alpha-2 country code"
          },
          "railsAllowed": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": ["ACH", "WIRE"]
            },
            "minItems": 1,
            "description": "Allowed payment rails"
          },
          "bankLast4": {
            "type": "string",
            "minLength": 4,
            "maxLength": 4,
            "description": "Last 4 digits of bank account"
          },
          "bankTokenHash": {
            "type": "string",
            "minLength": 1,
            "description": "Hash of encrypted bank token"
          },
          "version": {
            "type": "integer",
            "description": "Version number"
          },
          "status": {
            "type": "string",
            "enum": ["ACTIVE", "LOCKED", "PENDING_VERIFICATION"],
            "description": "Beneficiary status"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Creation timestamp"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "Last update timestamp"
          },
          "lastChangedAt": {
            "type": "string",
            "format": "date-time",
            "description": "Last change timestamp"
          },
          "lastChangedBy": {
            "type": "string",
            "description": "User ID who last changed the beneficiary"
          }
        }
      },
      "CreateBeneficiaryRequest": {
        "type": "object",
        "required": [
          "displayName",
          "country",
          "railsAllowed",
          "bankLast4",
          "bankTokenHash"
        ],
        "properties": {
          "displayName": {
            "type": "string",
            "minLength": 1,
            "description": "Display name"
          },
          "country": {
            "type": "string",
            "minLength": 2,
            "maxLength": 2,
            "description": "ISO 3166-1 alpha-2 country code"
          },
          "railsAllowed": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": ["ACH", "WIRE"]
            },
            "minItems": 1,
            "description": "Allowed payment rails"
          },
          "bankLast4": {
            "type": "string",
            "minLength": 4,
            "maxLength": 4,
            "description": "Last 4 digits of bank account"
          },
          "bankTokenHash": {
            "type": "string",
            "minLength": 1,
            "description": "Hash of encrypted bank token"
          }
        }
      },
...
    }
  }
}